added duel command, prevented commands sent by bots, fixed code in default case

+++ src/index.js
@@ -5,35 +5,56 @@ const prefix = '!';
 
 var err = 'error';
 
-/* notifies user when bot has turned on */
+/* 
+notifies user when bot has turned on 
+*/
 client.on('ready', function () {
     console.log('Bot started.');
     client.user.setActivity('Ready');
 }); // end start bot.on
 
-/* prints error */
+/* 
+prints error 
+*/
 function printError() {
-    console.log('An error has occured: ', err, '.');    
+    console.error('An error has occured: ', err, '.');    
 } // end printError
 
-/* allows user to enter commands for bot */
+/* 
+allows user to enter commands for bot 
+*/
 client.on('message', function (message) { 
+    // ignore commands bot sends
+    if (message.author.bot) return;
     if (!message.content.startsWith(prefix)) return;
     var args = message.content.substring(prefix.length).split(' ');
-    console.log(message.author.tag, ': ', message.content.toString());
+    author = message.guild.member(message.author);
+    console.log(author.nickname, ': ', message.content.toString());
     switch(args[0].toLowerCase()) {
-        /* prints hello world */
+        // prints hello world 
         case 'test':
             message.channel.send('Hello World!');
             break;
-        /* turns bot off */    
+        // starts a duel
+        case 'duel':
+            // gets the user first mentioned in the message
+            const user = message.mentions.users.first();
+            const member = message.guild.member(user);
+            if (args[1] == member) {    
+                message.channel.send(user + ' has been challenged!');
+
+                /* code for duel here */
+
+            }
+            else message.reply('usage: !duel <user>');
+            break;
+        // turns bot off     
         case 'quit':
             client.destroy();
             break;
-        /* gives message on how to use commands */    
+        // gives message on how to use commands    
         default:
-            err = 'unknown command';
-            message.channel.send('usage: !command');
+            message.channel.send('usage: !<command> <param>*');
     }
 }); // end command bot.on
